app/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
pydantic
yfinance